chore: rename local cache to server cache.
