+++ src/distribute_kv_cache/mod.rs
@@ -1,7 +1,7 @@
 //! Distributed cache module
 
 /// Local in-memory cache
-pub mod local_cache;
+pub mod server_cache;
 
 /// RPC module for cache
 pub mod rpc;
